Do not leak unknown asset paths into HTML
Fixes #3040
