+++ plugins/base/src/test/kotlin/resourceLinks/ResourceLinksTest.kt
@@ -17,9 +17,11 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
 import utils.TestOutputWriterPlugin
+import utils.assertContains
 import java.io.File
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
+import kotlin.test.assertTrue
 
 class ResourceLinksTest : BaseAbstractTest() {
     class TestResourcesAppenderPlugin(val resources: List<String>) : DokkaPlugin() {
@@ -202,6 +204,95 @@ class ResourceLinksTest : BaseAbstractTest() {
                             }
                         }
                 }
+            }
+        }
+    }
+
+    @Test
+    fun `should not add unknown resources to the end of the head section`() {
+        val baseConfiguration = DokkaBaseConfiguration(
+            customAssets = listOf(File("test/relativePath.js"))
+        )
+
+        val configuration = dokkaConfiguration {
+            sourceSets {
+                sourceSet {
+                    sourceRoots = listOf("src/main/kotlin")
+                }
+            }
+
+            pluginsConfigurations = mutableListOf(
+                PluginConfigurationImpl(
+                    fqPluginName = "org.jetbrains.dokka.base.DokkaBaseConfiguration",
+                    serializationFormat = DokkaConfiguration.SerializationFormat.JSON,
+                    values = toJsonString(baseConfiguration)
+                )
+            )
+        }
+
+        val writerPlugin = TestOutputWriterPlugin()
+        testInline(
+            """
+            |/src/main/kotlin/test/Test.kt
+            |package test
+            |
+            |class Test
+        """.trimMargin(),
+            configuration,
+            pluginOverrides = listOf(writerPlugin)
+        ) {
+            renderingStage = { _, _ ->
+                val generatedFiles = writerPlugin.writer.contents
+
+                val testPageHeadHtml = generatedFiles
+                    .getValue("root/test/-test/-test.html")
+                    .let { Jsoup.parse(it) }
+                    .select("head")
+                    .toString()
+
+                assertTrue {
+                    testPageHeadHtml.endsWith(
+                        """
+                         <script type="text/javascript" src="../../../scripts/symbol-parameters-wrapper_deferred.js" defer></script>
+                        </head>
+                        """.trimIndent()
+                    )
+                }
+            }
+        }
+    }
+
+    @Test
+    fun `should load script as defer if name ending in _deferred`() {
+        val configuration = dokkaConfiguration {
+            sourceSets {
+                sourceSet {
+                    sourceRoots = listOf("src/main/kotlin")
+                }
+            }
+        }
+
+        val writerPlugin = TestOutputWriterPlugin()
+        testInline(
+            """
+            |/src/main/kotlin/test/Test.kt
+            |package test
+            |
+            |class Test
+        """.trimMargin(),
+            configuration,
+            pluginOverrides = listOf(writerPlugin)
+        ) {
+            renderingStage = { _, _ ->
+                val generatedFiles = writerPlugin.writer.contents
+
+                assertContains(generatedFiles.keys, "scripts/symbol-parameters-wrapper_deferred.js")
+
+                val scripts = generatedFiles.getValue("root/test/-test/-test.html").let { Jsoup.parse(it) }.select("script")
+                val deferredScriptSources = scripts.filter { element -> element.hasAttr("defer") }.map { it.attr("src") }
+
+                // important to check symbol-parameters-wrapper_deferred specifically since it might break some features
+                assertContains(deferredScriptSources, "../../../scripts/symbol-parameters-wrapper_deferred.js")
             }
         }
     }
